refactor(validation): split zero-collection config coverage
Copilot-Session: 2209ac73-70c1-4a89-93b2-96ac59c0e91e
