+++ scripts/validation/check_zero_collection_tests.py
@@ -1,8 +1,4 @@
 #!/usr/bin/env python3
-# taste-lint: ignore file-size -- the guard, its config reader, its exemption
-# parser, and its skip-module trust check are one gate with one contract;
-# splitting them would let the pieces drift out of sync with what
-# tests/validation/test_check_zero_collection_tests.py exercises as a unit.
 """Fail when a file pytest walks into contributes no tests.
 
 A file that matches ``python_files`` inside ``testpaths`` is walked on every CI
